+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Head
 
 - Fixed: parsing `-` and `+` at the operation start in `operator-` rules.
+- Fixed: `findCommentsInRaws` false positives on comments inside strings (applicable to rules `double-slash-comment-inline`, `double-slash-comment-whitespace-inside`, `operator-no-unspaced`).
 
 # 1.3.3